improving cell detection and ocr by changing kernel values
In the original modules, the smallest cells were not detected because
they were smaller than the minimum threshold value for detection.
I solved this by decreasing the minimum threshold for detection.

Prior to ocr, the ocr_image module would clean the image of each cell.
This led to certain numbers being cropped and others deleted entirely.
I fixed this by increasing the kernel heigth and width for detecting
cell boundaries. Now numbers are not cleaned out.