Switched sequence of arguments of ocr_image.main to fit requirements of functools.partial
Functools.partial lets me freeze part of a function and arguments
so that I can use a partially preconfigured version of ocr_image
in the concurrent mapping function. I needed to switch the se-
quence of the arguments so that the static argument None comes first
and the iterable list of images comes second.