Reverted scale change in cell detection and added 5 pixels in table detection
The change of the scale variable in the cell detection had led to
many cells not being recognised properly. The 5 added pixels ensure
that the bottom of the tables is included in cropped table picture
in its entirety. When pictures were skewed, some parts of the table
bottom were left out of the cropped version, which led to failure
to detect bottom cells.