added check for number of unique columns to spot and remedy fused cells in cell extraction
For some tables, cell extraction with scale 5 resulted in horizontally fused cells
if they were small. Changing the scale to 6 helped with those tables, but led to other
errors with tables that worked fine with scale = 5. To remedy this, the cell extractor
now includes a check of the number of columns in the first ten rows. If there are more
than two different column numbers (indicating that in some of the rows, cells have
wrongly been fused, resulting in fewer columns), the cell extraction will run again,
this time with scale = 6.