Rewrote cell extraction to work with while loops to increase scale and included two checks
The cell extraction now includes two tests to decide whether a higher scale value should
be assigned: a test of whether there are different row lengths and a test of whether there
are rows present that have fewer than 13 columns. If either of these are now passed, then
the scale value is increased by 1. This is repeated until the tests are either passed or
the scale assumes a value of 10. Then a message is printed signalling that something is
still wrong with the cell extraction.