Added a version of pdf2images without rotation for working on manually rotated files and changed program for image extraction to pdftoppm
When working on the files that had been rotated manually, running the standard preprocessing on them results in them being rotated
falsely again. To circumvent this, I created a copy of the preprocessing code without the rotation function. This version is called
when going over the manually rotated files now.

I also changed the program that is called in a subprocess to extract the images from the pdf. I now use pdftoppm instead of pdf2images.
The latter did not handle the multiple elements in some of the files well. Pdftoppm does not have any problems and works well with the
rest of the code.