added check for failed table extraction
In some cases there is no extra space in an image to the left and right of the table.
In those cases, extending the cropping of the image to the table results in a failed
table extraction. To circumvent this, I added code to check whether the output of the
find_tables function is empty. If it is, the extra space around the table will be cut
iteratively until the function returns output that is not empty. If it is still empty
after having cut all the extra space, then a message is printed announcing this.