Unifying code variations in single scripts with if statements rather than having separate scripts
So far, different options in the pipeline were implemented by executing different scripts which varied
slightly from each other. This is not efficient and leads to redundant code. Now all options avaiable
in the scanner pipeline will be implemented in the original scripts. These scripts work with arguments
which trigger if statements.